+++ source/LibESE/src/D2Game/SKILLS/SkillSor_ESE.cpp
@@ -1062,7 +1062,7 @@ int32_t __fastcall ESE_SKILLS_EventFunc24_EnergyShield(D2GameStrc* pGame, int32_
     if (pUnit && nAbsorbedDamage > 0 && pSkillsTxtRecord->wPrgOverlay >= 0 && pSkillsTxtRecord->wPrgOverlay <= sgptDataTables->nOverlayTxtRecordCount)
     {
         int32_t nOverlayData[2] = {};
-        *(int64_t*)nOverlayData = D2Common_11053(UNITS_GetDirectionToCoords(pAttacker, CLIENTS_GetUnitX(pUnit), CLIENTS_GetUnitY(pUnit))) + pSkillsTxtRecord->wPrgOverlay;
+        *(int64_t*)nOverlayData = (int64_t)D2Common_11053(UNITS_GetDirectionToCoords(pAttacker, CLIENTS_GetUnitX(pUnit), CLIENTS_GetUnitY(pUnit))) + pSkillsTxtRecord->wPrgOverlay;
         UNITS_SetOverlay(pAttacker, nOverlayData[0], nOverlayData[1]);
     }
